move panic handlers to public helper macros
