+++ ui/helixflow-slint/Cargo.toml
@@ -4,11 +4,11 @@ version = "0.0.1"
 edition = "2024"
 
 [dependencies]
+assert_unordered.workspace = true
 helixflow-core.workspace = true
 slint.workspace = true
 
 [dev-dependencies]
-assert_unordered.workspace = true
 i-slint-backend-testing.workspace = true
 rstest.workspace = true
 uuid.workspace = true
